refactor(server): split ws.rs (1044 lines) into ws/ module
Concern-separated sub-modules for upgrade handshake, per-connection
session lifecycle, subscribe/unsubscribe handling, FPSS event broadcast,
contract bookkeeping, and outgoing-frame formatting. Pure move — zero
behavior change, full test parity.
